PR review feedback: rename variant & typecheck Continue Button CSS
